move standalone functions to utils
